move time formater to other class
