+++ app.rb
@@ -1,18 +1,12 @@
-require 'pry'
 
 class App
 
   def call(env)
-    perform_request
     [status, headers, body]
   end
 
   private
 
-
-  def perform_request
-    "request"
-  end
 
   def status
     200
