--- handler.rb
@@ -1,11 +1,0 @@
-#require 'rack'
-#
-#app = Proc.new do |env|
-#  [
-#    200,
-#    {'Content-Type' => 'text/plain'},
-#    ["Welcome abort!\n"]
-#  ]
-#end
-#
-#Rack::Handler::WEBrick.run app
